Rearrange file structure for routing
Introduce the 'routes' directory. Each .js file in this folder
is a separate page in the website. Index.js is the starting
point that manages all the routing and it also now contains
the firebase configuration. Also deleted unnecessary files,
including App.js, to simplify the project structure. Lastly,
began building the account creation screen.