Replace Login.js with SignIn.js using MUI
React's strength seems to be allowing devs to drop in
pre-built components. Material UI has many pre-built
components that fit the needs of our project and come
with nice styling. I redesigned the login screen with
MUI components and it took a fraction of the time and
it looks much better. I also began building the account
creation screen with a Radio component. Firebase
authentication still works.