Establish connection to firestore database
When an account is created, most of their information is now
stored in the firestore database in addition to their credentials
being added to the authentication system. Still unable to store DOB
info for now. I attempted to componentize the DOB picker but
it is not being used for now. Also added a logo to the top of
each screen.